refactor: replace api domain folders with flat HTTP function files
Delete 6 domain subdirectories (auth, projects, prs, chatSessions, fs, git)
and replace with flat files containing pure HTTP functions and type re-exports.
