--- frontend/src/shared/apis/auth/index.ts
@@ -1,3 +1,0 @@
-export { authQueryKey } from './queryKey';
-export { authQuery } from './query';
-export { authMutation } from './mutation';
