--- frontend/src/shared/apis/auth/mutation.ts
@@ -1,17 +1,0 @@
-import { mutationOptions } from '@tanstack/react-query';
-import { apiPost } from '../client';
-import type {
-  GitHubAuthStatus,
-  SwitchAccountBody,
-} from '@lgtmai/backend/types';
-
-export const authMutation = {
-  switchAccount: () =>
-    mutationOptions<GitHubAuthStatus, Error, SwitchAccountBody>({
-      mutationFn: (data) =>
-        apiPost<GitHubAuthStatus, SwitchAccountBody>(
-          '/api/auth/github/switch',
-          data
-        ),
-    }),
-};
