--- frontend/src/shared/apis/auth/queryKey.ts
@@ -1,5 +1,0 @@
-export const authQueryKey = {
-  all: ['auth'] as const,
-  githubStatuses: () => [...authQueryKey.all, 'githubStatus'] as const,
-  githubStatus: () => [...authQueryKey.githubStatuses()] as const,
-};
