--- frontend/src/shared/apis/chatSessions/index.ts
@@ -1,2 +1,0 @@
-export { chatSessionsQuery } from './query';
-export { chatSessionsQueryKey } from './queryKey';
