--- frontend/src/shared/apis/fs/index.ts
@@ -1,2 +1,0 @@
-export { fsQuery } from './query';
-export { fsQueryKey } from './queryKey';
