--- frontend/src/shared/apis/fs/queryKey.ts
@@ -1,5 +1,0 @@
-export const fsQueryKey = {
-  all: ['fs'] as const,
-  browses: () => [...fsQueryKey.all, 'browse'] as const,
-  browse: (path?: string) => [...fsQueryKey.browses(), path] as const,
-};
