--- frontend/src/shared/apis/git/index.ts
@@ -1,1 +1,0 @@
-export { gitMutation } from './mutation';
