--- frontend/src/shared/apis/projects/index.ts
@@ -1,3 +1,0 @@
-export { projectsQueryKey } from './queryKey';
-export { projectsQuery } from './query';
-export { projectsMutation } from './mutation';
