--- frontend/src/shared/apis/prs/index.ts
@@ -1,3 +1,0 @@
-export { prsQueryKey } from './queryKey';
-export { prsQuery } from './query';
-export { prsMutation } from './mutation';
